Test interaction for custom setting of parameters
Played around with FMOD/Unity to have a custom parameter. There is about 14 seconds of traffic noise, and the custom parameter "x" ranges from 0-25, with custom distortion automation spread across the whole thing. The parameter changes as time progresses. You can press K to stop the automation and have the sound progress to its end without the automation, or L to kill the sound